Removed unused files, project re-organized, added extension to easily instantiate a vc.
