--- banano-quest/banano-quest/Protocols/Banano.swift
@@ -1,20 +1,0 @@
-//
-//  Banano.swift
-//  banano-quest
-//
-//  Created by Pabel Nunez Landestoy on 6/28/18.
-//  Copyright © 2018 Michael O'Rourke. All rights reserved.
-//
-
-import Foundation
-import UIKit
-
-protocol Banano {
-    func refreshView()
-}
-
-class BananoQuestView: UIViewController, Banano {
-    func refreshView() {
-        // override
-    }
-}
