--- banano-quest/banano-quest/ViewController.swift
@@ -1,25 +1,0 @@
-//
-//  ViewController.swift
-//  banano-quest
-//
-//  Created by Michael O'Rourke on 6/19/18.
-//  Copyright © 2018 Michael O'Rourke. All rights reserved.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view, typically from a nib.
-    }
-
-    override func didReceiveMemoryWarning() {
-        super.didReceiveMemoryWarning()
-        // Dispose of any resources that can be recreated.
-    }
-
-
-}
-
