+++ banano-quest/banano-quest/ViewControllers/CompleteQuestViewController.swift
@@ -1,34 +1,42 @@
 //
-//  NewWalletViewController.swift
+//  CompleteQuestViewController.swift
 //  banano-quest
 //
-//  Created by Michael O'Rourke on 6/25/18.
+//  Created by Michael O'Rourke on 6/27/18.
 //  Copyright © 2018 Michael O'Rourke. All rights reserved.
 //
 
 import UIKit
+import MapKit
+import CoreLocation
 
-class NewWalletViewController: UIViewController {
+class CompleteQuestViewController: UIViewController {
 
+    @IBOutlet weak var mapView: MKMapView!
+    @IBOutlet weak var numberOfBananosLabel: UILabel!
+    @IBOutlet weak var bananoBackground: UIView!
+    
+    @IBOutlet weak var distanceLabel: UILabel!
+    @IBOutlet weak var prizeLabel: UILabel!
+    
     override func viewDidLoad() {
         super.viewDidLoad()
-
         // Do any additional setup after loading the view.
     }
-
+    
+    override func viewWillAppear(_ animated: Bool) {
+        super.viewWillAppear(animated)
+        
+        bananoBackground.layer.cornerRadius = bananoBackground.frame.size.width / 2
+        bananoBackground.clipsToBounds = true
+    }
+    
     override func didReceiveMemoryWarning() {
         super.didReceiveMemoryWarning()
         // Dispose of any resources that can be recreated.
     }
     
 
-    @IBAction func continuePressed(_ sender: Any) {
-        
-        let sb: UIStoryboard = UIStoryboard(name: "Questing", bundle: nil)
-        let vc = sb.instantiateViewController(withIdentifier: "QuestingVC") as? QuestingViewController
-        self.navigationController?.pushViewController(vc!, animated: true)
-        
-    }
     /*
     // MARK: - Navigation
