+++ banano-quest/banano-quest/ViewControllers/LandingViewController.swift
@@ -1,35 +1,29 @@
 //
-//  CompleteQuestViewController.swift
+//  LandingViewController.swift
 //  banano-quest
 //
-//  Created by Michael O'Rourke on 6/27/18.
+//  Created by Michael O'Rourke on 6/22/18.
 //  Copyright © 2018 Michael O'Rourke. All rights reserved.
 //
 
 import UIKit
-import MapKit
-import CoreLocation
 
-class CompleteQuestViewController: UIViewController {
+class LandingViewController: UIViewController {
 
-    @IBOutlet weak var mapView: MKMapView!
-    @IBOutlet weak var numberOfBananosLabel: UILabel!
-    @IBOutlet weak var bananoBackground: UIView!
-    
-    @IBOutlet weak var distanceLabel: UILabel!
-    @IBOutlet weak var prizeLabel: UILabel!
     override func viewDidLoad() {
         super.viewDidLoad()
-        bananoBackground.layer.cornerRadius = bananoBackground.frame.size.width / 2
-        bananoBackground.clipsToBounds = true
         // Do any additional setup after loading the view.
     }
     
+    override func viewWillAppear(_ animated: Bool) {
+        super.viewWillAppear(animated)
+        self.navigationController?.setNavigationBarHidden(true, animated: false)
+    }
+
     override func didReceiveMemoryWarning() {
         super.didReceiveMemoryWarning()
         // Dispose of any resources that can be recreated.
     }
-    
 
     /*
     // MARK: - Navigation
